requirements.txt
azure-cognitiveservices-speech==1.18.0
geocoder==1.38.1
geopy==2.2.0
pynput==1.8.3
pycountry==20.7.3
requests==2.26.0
unidecode==1.3.2
